Move SymDispatchMode to its own file
This is just code movement + a getter and a setter to break the
dependency of SymDispatchMode, and in turn, ProxySymDispatchMode on
sympy.

[ghstack-poisoned]
