+++ torch/fx/experimental/proxy_tensor.py
@@ -29,7 +29,8 @@ from torch.utils._python_dispatch import (
     _push_mode,
 )
 
-from .symbolic_shapes import ShapeEnv, SymDispatchMode, SymNode
+from .symbolic_shapes import ShapeEnv, SymNode
+from .sym_dispatch_mode import SymDispatchMode
 from torch.fx import Proxy
 import torch.fx.traceback as fx_traceback
 from torch import SymInt, SymFloat, SymBool
